// archive/musicml-langium/src/language-server/generated/ast.ts
export interface Bar extends AstNode {
    readonly $container: Track;
    readonly $type: 'Bar';
    notes: Array<Note>
    repeat?: number
}

export const Bar = 'Bar';

export function isBar(item: unknown): item is Bar {
    return reflection.isInstance(item, Bar);
}

export interface DrumNote extends AstNode {
    readonly $container: Note;
    readonly $type: 'DrumNote';
    notes: Array<string>
}

export const DrumNote = 'DrumNote';

export function isDrumNote(item: unknown): item is DrumNote {
    return reflection.isInstance(item, DrumNote);
}

export interface GlobalDirective extends AstNode {
    readonly $container: MusicPiece;
    readonly $type: 'GlobalDirective';
    midiClocksPerTick?: number
    notesPerMidiClock?: number
    tempo: number
    timeSignature: TimeSignature
}

export const GlobalDirective = 'GlobalDirective';

export function isGlobalDirective(item: unknown): item is GlobalDirective {
    return reflection.isInstance(item, GlobalDirective);
}

export interface MusicPiece extends AstNode {
    readonly $type: 'MusicPiece';
    composer?: string
    globalDirectives: GlobalDirective
    name: string
    tracks: Array<Track>
}

export const MusicPiece = 'MusicPiece';
// ...
export interface TimeSignature extends AstNode {
    readonly $container: GlobalDirective;
    readonly $type: 'TimeSignature';
    beatsPerBar: number
    noteValue: number
}

export const TimeSignature = 'TimeSignature';

export function isTimeSignature(item: unknown): item is TimeSignature {
    return reflection.isInstance(item, TimeSignature);
}

export interface Track extends AstNode {
    readonly $container: MusicPiece;
    readonly $type: 'Track';
    bars: Array<Bar>
    instrument: string
    name: string
    startTick?: number
}

export const Track = 'Track';